Fix detection of dodgy tags
Specifically, we were assuming the output had the format 'tag=value', and we
were looking for whitespace following the '='. However, it turns out that some
files have an '=' sign in their tag values, so we were flagging them by
mistake.

Now we cut the data at the first '=', and look for leading whitespace. That way,
any subsequent '=' will be ignored.