Use xvfb-run-safe when calling kid3-cli
Unfortunately kid3-cli requires a working X server to run, even though it
doesn't use it (it does offer clipboard integration, but we don't use it).

We were hard-coding DISPLAY to ':0' and hoping for the best. Unfortunately that
won't work in all situations, especially tests, and we want to make tests as
low-friction to write as possible!

I tried using xvfb-run, but that leaves crap around the filesystem, which causes
subsequent runs to break. I already solved this in warbo-utilities, so I've
resurrected the xvfb-run-safe script from there, and now use that in set_tag.